+++ src/EventListener/DoctrineEncryptListener.php
@@ -2,7 +2,6 @@
 
 namespace SpecShaper\EncryptBundle\EventListener;
 
-use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Event\OnFlushEventArgs;
 use Doctrine\ORM\Events;
 use Doctrine\Persistence\Event\LifecycleEventArgs;
@@ -13,6 +12,7 @@ use SpecShaper\EncryptBundle\BlindIndex\BlindIndexUpdater;
 use SpecShaper\EncryptBundle\Encryptors\EncryptorInterface;
 use SpecShaper\EncryptBundle\Exception\EncryptException;
 use Doctrine\Bundle\DoctrineBundle\Attribute\AsDoctrineListener;
+use SpecShaper\EncryptBundle\Mapping\EncryptedFieldMetadataProvider;
 
 /**
  * Doctrine event listener which encrypts/decrypts entities.
@@ -28,12 +28,6 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
     public const ENCRYPTOR_INTERFACE_NS = EncryptorInterface::class;
 
     /**
-     * An array of annotations which are to be encrypted.
-     * The default and initial is the bundle Encrypted Class.
-     */
-    protected array $annotationArray;
-
-    /**
      * Caches information on an entity's encrypted fields in an array keyed on
      * the entity's class name. The value will be a list of Reflected fields that are encrypted.
      */
@@ -41,21 +35,13 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
 
     private array $rawValues = [];
 
-    private bool $isDisabled;
-
-    /**
-     * @param EntityManagerInterface $em Deprecated in favour of fetching object manager from event args.
-     */
     public function __construct(
         private readonly EncryptorInterface $encryptor,
-        private readonly EntityManagerInterface $em,
-        array $annotationArray,
-        bool $isDisabled,
-        private readonly ?BlindIndexMetadataProvider $blindIndexMetadataProvider = null,
-        private readonly ?BlindIndexUpdater $blindIndexUpdater = null
+        private bool $isDisabled,
+        private readonly BlindIndexMetadataProvider $blindIndexMetadataProvider,
+        private readonly BlindIndexUpdater $blindIndexUpdater,
+        private readonly EncryptedFieldMetadataProvider $encryptedFieldMetadataProvider
     ) {
-        $this->annotationArray = $annotationArray;
-        $this->isDisabled = $isDisabled;
     }
 
     public function getEncryptor(): EncryptorInterface
@@ -143,7 +129,7 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
     {
         // Get the encrypted properties in the entity.
         $properties = $this->getEncryptedFields($objectManager, $entity);
-        $blindIndexes = $this->blindIndexMetadataProvider?->getForEntity($objectManager, $entity) ?? [];
+        $blindIndexes = $this->blindIndexMetadataProvider->getForEntity($objectManager, $entity);
 
         // If no encrypted properties, return false.
         if (empty($properties) && empty($blindIndexes)) {
@@ -157,10 +143,6 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
         $blindIndexesUpdated = false;
 
         if ($isEncryptOperation && !empty($blindIndexes)) {
-            if (null === $this->blindIndexUpdater) {
-                throw new EncryptException('Cannot create blind indexes; no blind index updater is configured.');
-            }
-
             $blindIndexesUpdated = $this->blindIndexUpdater->update($entity, $blindIndexes, $changeSet);
         }
 
@@ -245,23 +227,14 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
      */
     protected function getEncryptedFields(ObjectManager $objectManager, object $entity): array
     {
-        $reflectionClass = $this->getOriginalEntityReflection($objectManager, $entity);
-
-        $className = $reflectionClass->getName();
+        $classMetadata = $objectManager->getClassMetadata(get_class($entity));
+        $className = $classMetadata->getName();
 
         if (isset($this->encryptedFieldCache[$className])) {
             return $this->encryptedFieldCache[$className];
         }
 
-        $properties = $reflectionClass->getProperties();
-
-        $encryptedFields = [];
-
-        foreach ($properties as $key => $refProperty) {
-            if ($this->isEncryptedProperty($refProperty)) {
-                $encryptedFields[$key] = $refProperty;
-            }
-        }
+        $encryptedFields = $this->encryptedFieldMetadataProvider->getForClassMetadata($classMetadata);
 
         $this->encryptedFieldCache[$className] = $encryptedFields;
 
@@ -270,17 +243,7 @@ class DoctrineEncryptListener implements DoctrineEncryptListenerInterface
 
     private function isEncryptedProperty(ReflectionProperty $refProperty): bool
     {
-
-        // If PHP8, and has attributes.
-        if(method_exists($refProperty, 'getAttributes')) {
-            foreach ($refProperty->getAttributes() as $refAttribute) {
-                if (in_array($refAttribute->getName(), $this->annotationArray)) {
-                    return true;
-                }
-            }
-        }
-
-        return false;
+        return $this->encryptedFieldMetadataProvider->hasEncryptedAttribute($refProperty);
     }
 
     protected function getOriginalEntityReflection(ObjectManager $objectManager, $entity): \ReflectionClass